# Conflicts:
#	out/production/GF/gf/model/Membre.class
#	out/production/GF/gf/model/MembreFx.class
#	out/production/GF/gf/view/MembreDetailsController.class
#	out/production/GF/gf/view/MembrePanelController.class

+++ src/gf/model/Membre.java
@@ -7,14 +7,14 @@ public class Membre {
 
 	private String nom="";
 	private String prenom="";
-	private int telephone=0;
-	private int cni=0;
+	private long telephone=0;
+	private long cni=0;
 	private String adresse="";
 	private String photo="";
 	private long id = -1;
 
 		
-	public Membre(String nom, String prenom, int telephone, int cni, String adresse, String photo){
+	public Membre(String nom, String prenom, long telephone, long cni, String adresse, String photo){
 		this.nom=nom;
 		this.prenom=prenom;
 		this.telephone=telephone;
@@ -23,7 +23,7 @@ public class Membre {
 		this.photo=photo;
 	}
 
-	public Membre(long id, String nom, String prenom, int telephone, int cni, String adresse, String photo){
+	public Membre(long id, String nom, String prenom, long telephone, long cni, String adresse, String photo){
 		this.id=id;
 		this.nom=nom;
 		this.prenom=prenom;
